added questions to back end
